PLAT-9655 Refactor Webhooks to Enable updating a webhooks from settings and refator callback to accomodate for query changes
In This Commint / code change  we are refactoring the webhhok to accommodate for the PHPSDK update in query builder method also to make the webhook update once the settings are saved in order to accommodate for any changes brought to the query saved in those webhooks config.
Also to change to payload sent back to the website from being a full Course / LP GraphQL response on the course or LP to just returning the ID of that Course and LP and the website sending a call to the TMS Core API to fetch the course content to synch this way any change we introduce to Query fetching the courser or LP will be immediately reflected in the webhooks, also this change will pave the ground to the next enhancement of buffering the Callbacks incase many webhook call backs hits the website at once (or from a bulk update in the TMS).
https://administrate.atlassian.net/browse/PLAT-9655