PS-3486 Update Search filter using weblink-api to fetch and save the Categories as filter options.
In this commit we made the search category filter display categories for courses from weblink-api the same way as the location filter dose.
This will make sure that the categories visible in the search are similar to the one available for the configured portal in the plugin settings page.
Prior to this update the Categories where fetched and displayed from wordpress synched categories.