PLAT-15138 Add custom filter admwpp_course_post_status to expose capabilty to alter course post status
In this commit we are adding a new custom filter admwpp_course_post_status to enable us / devs to alter the course post status based on synched most meta values to WordPress.
Example: For Maersk, we need to be able to hide the LPs that have isBundle set to true so on the theme level we can use the filter to override the post_status and make all the LPs with is a bundle set to true have the status as pending and hidden from the public as it's needed for Maersk, also this will propagate to the views created for the title suggestions.
The usage of this filter can extend to other conditions to set the post_status as needed like TMS custom filed values that are also synched to WordPress such as the example show_in_catalogue custom field. (CGA / Maersk)