NI-4953 Change part of day afternoon and evening endtime
We had to change the part of day afternoon and evening endtime value and update it with 1 min overlap.
Basically the courses that have events starting at 14:00 and ending at 17:00 don't appear in the results when a user filters by part of day
and selects afternoon, and the main reason for that is that the end time field is being set to 17:00:59 instead of 16:59:59.